# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  utils.Response:
    properties:
      message:
        type: string
      pageInfo:
        additionalProperties: {}
        type: object
      result: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: This is a CRUD API for managing users.
  version: "1.0"
paths:
  /users:
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, and password
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User created successfully
          schema:
            $ref: '#/definitions/utils.Response'
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/utils.Response'
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user details by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User detail retrieved successfully
          schema:
            $ref: '#/definitions/utils.Response'
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update a user by ID (partial update)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User object to be updated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/utils.Response'
      tags:
      - Users
  /users/list-all:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: List of all users
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all users
      tags:
      - Users
swagger: "2.0"